Optimise transaction script
Replace HASH256 with HASH160 in the TX1/TX3 transaction script, to match the hashing cycles normally used on public keys, and reduce space required for the script.
Restructure TX1/TX3 transaction script to move the common public key verification step outside of the conditional section, removing a point of duplication.